Pushing new chnages to tooltip
