+++ package.json
@@ -36,6 +36,7 @@
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
     "react-native-svg": "13.9.0",
+    "react-native-uuid": "^2.0.1",
     "react-native-webview": "13.2.2",
     "react-redux": "^9.0.4",
     "reanimated-color-picker": "^3.0.3",
